Implementation of reason for closing and closing message being supplied with reason
